pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-classic>=1.0.0",
    "langchain-google-genai>=4.1.2",
    "langchain-mongodb>=0.9.0",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-mongodb>=0.3.0",
    "langgraph-checkpoint-postgres>=3.0.2",
    "langgraph-store-mongodb>=0.1.1",
    "langmem>=0.0.30",
    "psycopg[binary]>=3.3.2",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]
